+++ packages/nocodb-sdk/src/lib/columnHelper/columns/Currency.ts
@@ -1,6 +1,4 @@
-import { AbstractColumnHelper } from '..';
-import { parsePercentValue } from '../utils/parser';
-import { serializePercentValue } from '../utils/serializer';
+import { AbstractColumnHelper, parsePercentValue, serializePercentValue } from '..';
 
 export class CurrencyHelper extends AbstractColumnHelper {
   columnDefaultMeta = {
