+++ packages/nocodb-sdk/src/lib/index.ts
@@ -51,3 +51,4 @@ export * from '~/lib/import-export-data';
 export * from '~/lib/is';
 export * from '~/lib/durationUtils';
 export * from '~/lib/sync';
+export * from '~/lib/columnHelper';
